Fix add payment amount if customer has paid

+++ database/migrations/2020_08_04_204207_create_orders_table.php
@@ -18,7 +18,7 @@ class CreateOrdersTable extends Migration
             $table->timestamps();
             $table->float('price');
             $table->boolean('has_paid')->default(false);
-            $table->integer('payment_amount')->nullable();
+            $table->float('payment_amount')->nullable();
             $table->string('notes')->nullable();
 			$table->string('status')->nullable();
 			$table->softDeletes();
